requirements/base.txt
# requirements/base.txt
Django==5.0.1
python-dotenv==1.0.0
django-allauth==0.60.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-ckeditor==6.7.0
django-compressor==4.4
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-storages==1.14.2
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
Pillow==11.2.1